Refactored nav stuff into it's own module, not lazy loaded, but self contained.
