+++ UI/Web/src/app/nav/events-widget/events-widget.component.scss
