+++ UI/Web/src/app/nav/grouped-typeahead/grouped-typeahead.component.scss
