+++ UI/Web/src/app/nav/nav-header/nav-header.component.scss
